# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Service Portal API",
    "version": "1.0.0",
...
  },
  "paths": {
...
    "/api/v1/statuses": {
      "get": {
        "tags": [
          "statuses"
        ],
        "summary": "List the service request statuses",
        "description": "Every seeded status, in display order (ST-1, ST-2).\n\nOrdered by ``sort_order`` in SQL, not by insertion or by name: the column\nexists precisely so the stepper's order is data rather than a convention\nthe frontend has to re-know. ``open`` before ``closed`` is not\nalphabetical, and rows come back in no guaranteed order without an\n``ORDER BY``.",
        "operationId": "list_statuses_api_v1_statuses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusList"
                }
              }
            }
          },
          "500": {
            "description": "Returned when the request failed for a reason the API does not attribute to the caller — `message` never carries internal detail and the traceback is logged server-side (XC-14).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service-requests": {
      "get": {
        "tags": [
          "service-requests"
        ],
        "summary": "List service requests visible to the caller",
        "description": "A page of requests the caller may see (SR-1 to SR-5, SR-15, XC-10, XC-11).",
        "operationId": "list_service_requests_api_v1_service_requests_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "low",
                    "medium",
                    "high"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ServiceRequestOut_"
                }
              }
            }
          },
          "422": {
            "description": "Returned when one or more supplied fields are invalid (XC-4).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Returned when the `access_token` cookie is absent, expired, or no longer names an active user — no distinction is drawn between those cases (XC-5).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Returned when the request failed for a reason the API does not attribute to the caller — `message` never carries internal detail and the traceback is logged server-side (XC-14).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessTokenCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "service-requests"
        ],
        "summary": "File a service request",
        "description": "File a request on behalf of the caller (SR-6 to SR-10, SR-14).\n\nThree of the row's columns are never read from the body (SR-10 / XC-8):\n``requestor_id`` is the authenticated user, ``request_type`` is\n``'general'``, and ``current_status_id`` is whatever ``'open'``'s id is.\nThe schema has no fields for them, so a body supplying all three is\naccepted and creates exactly the same row as one that doesn't — which is\nXC-8's actual requirement, and the reason the schema ignores unknown keys\nrather than rejecting them.",
        "operationId": "create_service_request_api_v1_service_requests_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRequestCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Returned when one or more supplied fields are invalid (XC-4).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Returned when the `access_token` cookie is absent, expired, or no longer names an active user — no distinction is drawn between those cases (XC-5).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Returned when the `X-Requested-With` header is missing, which middleware refuses before routing (XC-9).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Returned when the request failed for a reason the API does not attribute to the caller — `message` never carries internal detail and the traceback is logged server-side (XC-14).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessTokenCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Requested-With",
              "default": "XMLHttpRequest"
            },
            "description": "CSRF guard (XC-9). Any non-empty value. Enforced by middleware before routing, so a request without it is answered `403` without its body or its cookies being read — which is why it is absent from this operation's handler signature."
          }
        ]
      }
    },
    "/api/v1/service-requests/{request_id}": {
      "get": {
        "tags": [
          "service-requests"
        ],
        "summary": "Fetch one service request",
        "description": "One request, if the caller may see it (SR-11 to SR-13).\n\n``request_id`` is typed ``str`` deliberately — see ``parse_uuid_or_none``.",
        "operationId": "get_service_request_api_v1_service_requests__request_id__get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestOut"
                }
              }
            }
          },
          "401": {
            "description": "Returned when the `access_token` cookie is absent, expired, or no longer names an active user — no distinction is drawn between those cases (XC-5).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Returned when no service request with that id is visible to the caller — a malformed id, an id matching no row, and an id belonging to someone else are answered identically by design (XC-7, SR-13).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Returned when the request failed for a reason the API does not attribute to the caller — `message` never carries internal detail and the traceback is logged server-side (XC-14).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessTokenCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Page_ServiceRequestOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ServiceRequestOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[ServiceRequestOut]"
      },
...
      "ServiceRequestCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Description"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Priority"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "priority"
        ],
        "title": "ServiceRequestCreate",
        "description": "``POST /service-requests`` body (SR-6 through SR-10)."
      },
      "ServiceRequestOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "request_type": {
            "type": "string",
            "title": "Request Type"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "status": {
            "$ref": "#/components/schemas/StatusOut"
          },
          "requestor": {
            "$ref": "#/components/schemas/UserSummary"
          },
          "assignee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "request_type",
          "priority",
          "status",
          "requestor",
          "assignee",
          "created_at",
          "updated_at",
          "description"
        ],
        "title": "ServiceRequestOut",
        "description": "The one ``ServiceRequest`` shape, for list and detail alike.\n\ndesign.md §0 decision 2: not two shapes. ``description`` is present in both\n(SR-5) — one text field per row is cheap next to having two payloads that\nare almost the same, which is the inconsistency this contract exists to\nresolve.\n\nFields are ordered as design.md §4 writes them."
      },
...
      "StatusList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StatusOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "StatusList",
        "description": "``GET /statuses``' body (ST-2).\n\nA bare ``{\"items\": [...]}``, deliberately **not** XC-10's paginated\nenvelope: the table holds exactly four seeded rows, always, so ``total`` /\n``page`` / ``page_size`` would be constants dressed up as data. A separate\nmodel rather than a raw dict so the shape reaches the OpenAPI document."
      },
      "StatusOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "is_terminal": {
            "type": "boolean",
            "title": "Is Terminal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "sort_order",
          "is_terminal"
        ],
        "title": "StatusOut",
        "description": "One row of the ``statuses`` reference table (design.md §1)."
      },
      "UserSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "role"
        ],
        "title": "UserSummary",
        "description": "A user reference embedded in another resource (design.md §1)."
      },
      "ErrorDetail": {
        "title": "ErrorDetail",
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "title": "Code",
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "VALIDATION_ERROR",
              "INTERNAL_ERROR"
            ],
            "description": "Machine-readable error class. Closed set."
          },
          "message": {
            "title": "Message",
            "type": "string",
            "description": "Human-readable summary, safe to show a user. Never carries server internals."
          },
          "fields": {
            "title": "Fields",
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Per-field messages, keyed by field name. Present only on `VALIDATION_ERROR`."
          }
        }
      },
      "ErrorEnvelope": {
        "title": "ErrorEnvelope",
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "description": "The one shape every error response in this API uses."
      }
    },
    "securitySchemes": {
      "accessTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "Session access token. Set by `POST /api/v1/auth/login` and `POST /api/v1/auth/refresh` as an `httpOnly`, `Secure`, `SameSite=Lax` cookie with a 1 hour lifetime. Sent automatically by the browser; not readable from JavaScript and not settable by a client."
      },
      "refreshTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "Opaque refresh token, 30 day lifetime, same cookie attributes as `access_token`. Read by `POST /api/v1/auth/refresh` only, which rotates it; replaying an already-rotated token revokes the whole family (AUTH-11)."
      }
    }
  }
}